- Moved append_flow and sink writers to pipeline writer file - Added unit tests
